+++ src/components/Elements/Breadcrumbs/BreadcrumbMenu.js
@@ -49,4 +49,4 @@ BreadcrumbMenu.propTypes = {
 };
 
 
-export default BreadcrumbMenu;
+export default React.memo(BreadcrumbMenu);